fix location string bug firefox incompatibility
`origin` is different in chrome than in firefox. So when the the
new location string was formed, in chrome I used the `origin` to form
the "file://" protocol part of the string. This doesn't work in firefox,
because firefox has "null" as the origin. What I do now is use the
`protocol` which is "file:" for both chrome and firefox, and then I just
add a couple of forward slashes to get the same effect as before that I
had when in chrome.